requirements.txt
# Core dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.31

# Security & PII Detection  
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# LLM Providers
requests==2.32.3
httpx==0.27.0
aiohttp==3.9.5
google-generativeai==0.5.4
openai==1.35.3

# Utilities
python-multipart==0.0.9
pyyaml==6.0.1
python-dotenv==1.0.1
cryptography==42.0.8

# Logging
colorlog==6.8.2